requirements.txt
Django==2.2.6
djangorestframework==3.11.0
djoser==2.0.3
requests==2.18.4

